kraken-wrapper/: hygiene: Remove abstraction over build environments, there is only one `BuildEnv` implementation now and that is Uv
