+++ kraken-wrapper/src/uv/__main__.pyi
@@ -1,0 +1,1 @@
+def find_uv_bin() -> str: ...
